uqmi: improve response detection
Create a helper which checks the flags depending on the service
indicated in the received message, as the response flag is not the same
for the CTL service and the other services. This avoids considering
a CTL indication as a valid response.

Also use a mask in order to check the flags, as they are a bitmap.

Signed-off-by: Jean Thomas <jean.thomas@wifirst.fr>